% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_analysisModule.R
\name{annoCell}
\alias{annoCell}
\title{Single-Cell Type Annotation via annoCell (SingleR-based)}
\usage{
annoCell(
  seurat.object,
  method = "SingleR",
  clustered = FALSE,
  host_organism = "human",
  custom_ref = NULL,
  ...
)
}
\arguments{
\item{seurat.object}{seurat object.}

\item{method}{Default: \code{"SingleR"}. Cell type annotation method.}

\item{clustered}{Default: \code{FALSE}.
User could set it with \code{TRUE} if \code{\link[Seurat]{FindClusters}}
function has been done on the seurat object.}

\item{host_organism}{Default: \code{"human"}.
Organism of host single cell RNA-seq data,
to determine which reference to use.
Selected from \code{["huamn", "mouse"]}.
User could also use other reference by using param \code{"custom_ref"}.
\describe{
\item{\code{"huamn"}}{Using \code{"HumanPrimaryCellAtlasData()"} as reference.
Details see \code{\link[celldex]{HumanPrimaryCellAtlasData}}}
\item{\code{"mouse"}}{Using \code{"MouseRNAseqData()"} as reference.
Details see \code{\link[celldex]{MouseRNAseqData}}}
}}

\item{custom_ref}{Default: \code{NULL}. \cr
\strong{- \code{\link{SingleR}} bulid-in references, have been migrated to the \code{celldex} package.}
\describe{
\item{\code{"HumanPrimaryCellAtlasData()"}}{\code{\link[celldex]{HumanPrimaryCellAtlasData}}.
Normalized expression values for 713 microarray samples from the Human Primary Cell Atlas}
\item{"MouseRNAseqData()"}{\code{\link[celldex]{MouseRNAseqData}}.
Normalized expression values of 358 bulk RNA-seq samples of sorted cell populations that can be found at GEO.}
\item{\code{"ImmGenData()"}}{\code{\link[celldex]{ImmGenData}}.
Normalized expression values of 830 microarray samples of pure mouse immune cells, generated by the Immunologic Genome Project (ImmGen).}
\item{...}{...}
}
\strong{- Additional references. Details see chapter 3 in \code{\link{SingleR}} package document.}}

\item{...}{additional params from function \strong{SingleR}.
Details params in \code{\link[SingleR]{SingleR}}.}
}
\value{
seurat object
}
\description{
Single-Cell Type Annotation via annoCell (SingleR-based)
}
\examples{
# data("seurat.object")
# seurat.object <- runSeurat(seurat.object)
# seurat.object <- annoCell(seurat.object)

}
