% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{constructPhyloseq}
\alias{constructPhyloseq}
\title{Convert mpa format abundance matrix to phyloseq object}
\usage{
constructPhyloseq(microbe.counts)
}
\arguments{
\item{microbe.counts}{mpa format abundance matrix}
}
\value{
phyloseq object
}
\description{
Convert mpa format abundance matrix to phyloseq object
}
