% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_classes.R
\docType{class}
\name{corrResult-class}
\alias{corrResult-class}
\title{corrResult Class}
\description{
A class to hold correlation and p-values matrices.

The `corrResult` class is designed to store the results of a correlation analysis, including
the correlation matrix, p-values matrix, and any parameters used during the analysis.
}
\section{Slots}{

\describe{
\item{\code{correlation_matrix}}{A matrix or data frame containing correlation values.}

\item{\code{p_values_matrix}}{A matrix or data frame containing p-values associated with the correlation values.}

\item{\code{parameters}}{A list of parameters used in the analysis of correlation.}
}}

