% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{mpa2tax}
\alias{mpa2tax}
\title{Convert mpa format object to taxonomy table}
\usage{
mpa2tax(taxonomy_list)
}
\arguments{
\item{taxonomy_list}{String format vector}
}
\value{
taxonomy table
}
\description{
Convert mpa format object to taxonomy table
}
