% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{phyloseq_select_level}
\alias{phyloseq_select_level}
\title{Select Microbial Features by Taxonomic Level from a phyloseq Object.}
\usage{
phyloseq_select_level(microbe.phyloseq.object, tax_level = "genus")
}
\arguments{
\item{microbe.phyloseq.object}{phyloseq object}

\item{tax_level}{Default: \code{"genus"}}
}
\value{
microbe phyloseq object
}
\description{
Select Microbial Features by Taxonomic Level from a phyloseq Object.
}
