% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{phyloseq_select_tax}
\alias{phyloseq_select_tax}
\title{select abundance of tax_name at tax_level level.}
\usage{
phyloseq_select_tax(
  phyloseq.object,
  tax_name = "Bacillus",
  tax_level = "genus"
)
}
\arguments{
\item{phyloseq.object}{phyloseq object}

\item{tax_name}{Default: \code{"Bacillus"}}

\item{tax_level}{Default: \code{"genus"}}
}
\value{
phyloseq object at selected tax level.
}
\description{
select abundance of tax_name at tax_level level.
}
\examples{
NULL
}
