% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotAlpha}
\alias{plotAlpha}
\title{Visual Representation of Intracellular Microbiome Diversity.}
\usage{
plotAlpha(
  seurat.object,
  alpha.index = "Simpson",
  tax_level = "species",
  group.by = "orig.ident",
  reduction = NULL,
  custom.color = NULL
)
}
\arguments{
\item{seurat.object}{seurat object after running \code{\link{runSeurat}} and \code{\link{calcAlpha}}}

\item{alpha.index}{Default: \code{"Simpson"},
string or vector, alpha diversity index name.}

\item{tax_level}{Default: \code{"species"} taxonomy level.}

\item{group.by}{Default: \code{"orig.ident"},
Name of seurat object metadata columns to group (color) cells by.}

\item{reduction}{Default: \code{NULL}.
Which dimensionality reduction to use. Selected from \code{"umap"},  \code{"tsne"}, or \code{"pca"}.}

\item{custom.color}{Default: \code{NULL}, which will use 20 built-in colors \code{c('#D97894','#7C98CE','#BBCB54', '#A7BED3',
'#C9A2E1', '#748977', '#BEBE9E', '#4CD28F',
'#CFAB84', '#E3E15C', '#DFC7CF', '#6BB8BF',
'#CBCBCB', '#9696eb', '#DED1AD', '#E29F58',
'#D8A1CE', '#B16455', '#ad9bf4', '#BEB68D')}.
Users could provide a vector of colors, each color corresponds to one group mapped to param \code{"group.by"}}
}
\value{
ggplot2 format graph
}
\description{
Visual Representation of Intracellular Microbiome Diversity.
}
\examples{
data("seurat.object")
# seurat.object <- runSeurat(seurat.object)
seurat.object <- calcAlpha(seurat.object)
plotAlpha(seurat.object)
# plotAlpha(seurat.object, alpha.index = "Shannon", reduction = "umap")
}
