% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotCore}
\alias{plotCore}
\title{Core Microbiome Visualization Based on Prevalence and Abundance Thresholds.}
\usage{
plotCore(
  seurat.object,
  tax_level = "species",
  group.by = "orig.ident",
  color = c("#EFEFEF", "#A50026")
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{tax_level}{Default: \code{"species"}. Taxonomy level.}

\item{group.by}{Default: \code{"orig.ident"}.
Name of seurat object metadata columns to group (color) cells by.}

\item{color}{Default: \code{c("white", "red")}.
Colors of the core microbe heatmap.
User could provide a vector of colours to use for n-colour gradient.
E.g: \code{"colorRampPalette(c("white", "red"))(256)"}}
}
\value{
heatmap
}
\description{
Core Microbiome Visualization Based on Prevalence and Abundance Thresholds.
}
\examples{
data("seurat.object")
plotCore(seurat.object)

}
