% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Visualize Feature Abundance or Expression Across Samples or Groups.}
\usage{
plotFeatures(
  seurat.object,
  tax_name = NULL,
  tax_level = "species",
  features = c("nCount_RNA", "nFeature_RNA", "nCount_Microbe", "nFeature_Microbe"),
  reduction = "umap",
  ncol = 2,
  nrow = 2,
  labels = NULL,
  spt.sc.theme = FALSE,
  ...
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{tax_name}{Default: \code{NULL}. Microbe tax_name from different taxonomy levels
(e.g. [\code{"Escherichia_coli", tax_level = "species"}],
[\code{"Bacillota", tax_level = "phylum"]}.)}

\item{tax_level}{Default: \code{"species"}. Taxonomy level of param \strong{tax_name}.}

\item{features}{Default: \code{c("nCount_RNA", "nFeature_RNA",
"nCount_Microbe", "nFeature_Microbe")}.
Vector of features to plot.
Details see \strong{features} in \code{\link[Seurat]{FeaturePlot}}.\cr
Features can come from:
\itemize{
  \item feature (e.g. a gene name - "MS4A1")
  \item A column name from meta.data (e.g. mitochondrial percentage - \code{"percent.mito"}, \code{"nCount_Microbe"})
  \item A column name from a DimReduc object corresponding to the cell embedding values (e.g. the PC 1 scores - "PC_1")
}}

\item{reduction}{Defalut: \code{"umap"}. Details see \strong{reduction} in \code{\link[Seurat]{FeaturePlot}}.
Which dimensionality reduction to use.
If not specified, first searches for umap, then tsne, then pca.}

\item{ncol}{Default: \code{2}.
(optional) Number of rows in the plot grid.
Details see \strong{ncol} in \code{\link[cowplot]{plot_grid}}.}

\item{nrow}{Default: \code{2}.
(optional) Number of columns in the plot grid.
Details see \strong{nrow} in \code{\link[cowplot]{plot_grid}}.}

\item{labels}{Default: \code{NULL}.
(optional) List of labels to be added to the plots.
Details see \strong{labels} in \code{\link[cowplot]{plot_grid}}.
\describe{
\item{\code{"AUTO"}}{auto-generate upper-case labels, "A", "B"...}
\item{\code{"auto"}}{auto-generate lower-case labels, "a", "b"...}
}}

\item{spt.sc.theme}{Default: \code{FALSE}.
If the data is spatial transcriptomics,
should the visualization theme for single-cell sequencing data analysis be used?}

\item{...}{additional params from function \strong{FeaturePlot}.
Details params in \code{\link[Seurat]{FeaturePlot}}.}
}
\value{
feature plot
}
\description{
Visualize Feature Abundance or Expression Across Samples or Groups.
}
\examples{
data("seurat.object")
seurat.object <- runSeurat(seurat.object)
plotFeatures(seurat.object, labels = "auto")
# plotFeatures(seurat.object, features = c("PLA2G2A", "CDA", "HSPG2"), labels = "AUTO")
# plotFeatures(seurat.object, tax_name = "Escherichia_coli")
# plotFeatures(seurat.object, tax_name = "Bacillota", tax_level = "phylum")

}
