% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap Visualization of Differentially Expressed Marker Genes.}
\usage{
plotHeatmap(
  seurat.object,
  dtype = "host.marker",
  row.group.by = "seurat_clusters",
  col.cluster.by = "seurat_clusters",
  feature.row = 3
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{dtype}{Default: \code{"host.marker"}.}

\item{row.group.by}{Default: \code{"seurat_clusters"}.}

\item{col.cluster.by}{Default: \code{"seurat_clusters"}.}

\item{feature.row}{Default: \code{3}.}
}
\value{
heatmap plot
}
\description{
Heatmap Visualization of Differentially Expressed Marker Genes.
}
\examples{
#
}
