% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotStackedBar}
\alias{plotStackedBar}
\title{Visualize Taxonomic Composition Across Samples Using Stacked Bar Charts.}
\usage{
plotStackedBar(
  seurat.object,
  tax_level = "species",
  group.by = "orig.ident",
  custom.color = NULL,
  cluster_group = FALSE,
  top_n = 20
)
}
\arguments{
\item{seurat.object}{seurat object.}

\item{tax_level}{Default: \code{"species"}. Taxonomy level of param \strong{feature}.}

\item{group.by}{Default: \code{"orig.ident"}.
Name of seurat object metadata columns to group (color) cells by.}

\item{custom.color}{Default: \code{NULL}. which will use 20 built-in colors \code{c('#D97894','#7C98CE','#BBCB54', '#A7BED3',
'#C9A2E1', '#748977', '#BEBE9E', '#4CD28F',
'#CFAB84', '#E3E15C', '#DFC7CF', '#6BB8BF',
'#CBCBCB', '#9696eb', '#DED1AD', '#E29F58',
'#D8A1CE', '#B16455', '#ad9bf4', '#BEB68D')}.
Users could provide a vector of colors, each color corresponds to one group mapped to param \code{"group.by"}.}

\item{cluster_group}{Default: \code{FALSE}. Weather cluster the groups with a tree based on bray distance.}

\item{top_n}{Default: \code{20}. Numbers of enriched taxonomy to show.}
}
\value{
Stark Bar
}
\description{
Visualize Taxonomic Composition Across Samples Using Stacked Bar Charts.
}
\examples{
data("seurat.object")
plotStackedBar(seurat.object)
# seurat.object <- runSeurat(seurat.object)
# plotStackedBar(seurat.object, group.by = "seurat_clusters", tax_level = "phylum")

}
