% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Trajectory Visualization for Host Transcriptomes and Microbial Communities.}
\usage{
plotTrajectory(
  seurat.object,
  method = "monocle3",
  feature = "pseudotime",
  gene_name = NULL,
  tax_name = NULL,
  tax_level = "species",
  custom.color = NULL,
  ...
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{method}{Default: \code{"monocle3"}.}

\item{feature}{Default: \code{"pseudotime"}.}

\item{gene_name}{Default: \code{NULL}.}

\item{tax_name}{Default: \code{NULL}.}

\item{tax_level}{Default: \code{"species"}.}

\item{custom.color}{Default: \code{NULL}.}

\item{...}{additional params from function \strong{plot_cells}.
Details params in \code{\link[monocle3]{plot_cells}}.}
}
\value{
Trajectory plot.
}
\description{
Trajectory Visualization for Host Transcriptomes and Microbial Communities.
}
\examples{
# data("seurat.object")
# seurat.object <- runSeurat(seurat.object)
# seurat.object <- runTrajectory(seurat.object)
# plotTrajectory(seurat.object, feature = "State")
# plotTrajectory(seurat.object, gene_name = "IGHM")
# plotTrajectory(seurat.object, tax_name = "Escherichia", tax_level = "genus")

}
