% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_plotModule.R
\name{plotUpset}
\alias{plotUpset}
\title{UpSet Plot for Shared and Unique Microbial Features Across Groups.}
\usage{
plotUpset(
  seurat.object,
  tax_level = "species",
  group.by = "orig.ident",
  custom.color = NULL
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{tax_level}{Default: \code{"species"}. Taxonomy level.}

\item{group.by}{Default: \code{"orig.ident"}.
Name of seurat object metadata columns to group (color) cells by.}

\item{custom.color}{Default: \code{NULL}. which will use 20 built-in colors \code{c('#D97894','#7C98CE','#BBCB54', '#A7BED3',
'#C9A2E1', '#748977', '#BEBE9E', '#4CD28F',
'#CFAB84', '#E3E15C', '#DFC7CF', '#6BB8BF',
'#CBCBCB', '#9696eb', '#DED1AD', '#E29F58',
'#D8A1CE', '#B16455', '#ad9bf4', '#BEB68D')}.
Users could provide a vector of colors, each color corresponds to one group mapped to param \code{"group.by"}.}
}
\value{
upset plot
}
\description{
UpSet Plot for Shared and Unique Microbial Features Across Groups.
}
\examples{
plotUpset(seurat.object)

}
