% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_readData.R
\name{read_Host}
\alias{read_Host}
\title{Load Host Single-Cell or Spatial Transcriptomic Data.}
\usage{
read_Host(named.outs.dirs, data.type = "sc")
}
\arguments{
\item{named.outs.dirs}{Default: String or string vector, representing the paths to the outs directories of the output results from Cell Ranger or Spacer Ranger.}

\item{data.type}{Default: \code{"sc"}. Selected from \code{"sc"} and \code{"spt"}, sc represents single cell transcriptome data;
spt represents spatial transcriptome sequencing data.}
}
\value{
seurat object
}
\description{
Load Host Single-Cell or Spatial Transcriptomic Data.
}
\examples{
named.outs.dirs <- c(
  system.file("extdata",
    "sc/SAMN20336351/outs/",
    package = "intraMisc"
  ),
  system.file("extdata",
    "sc/SAMN20336352/outs/",
    package = "intraMisc"
  )
)
names(named.outs.dirs) <- c("SAMN20336351", "SAMN20336352")
seurat.object <- read_Host(named.outs.dirs)
}
