% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_analysisModule.R
\name{runSeurat}
\alias{runSeurat}
\title{Basic Processing Workflow for Seurat Objects.}
\usage{
runSeurat(
  seurat.object,
  filter.cells = TRUE,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  nfeatures = 2000,
  integrate.method = Seurat::CCAIntegration,
  integrate.new.reduction = "cca",
  dim.run.umap = 1:20,
  dim.run.tsne = 1:20,
  dim.find.neighbors = 1:20,
  resolution.find.clusters = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{filter.cells}{Default: \code{FALSE}. Logical.
Weather to filtering cells by the percentage of "mitochondrial genes".}

\item{normalization.method}{Default: \code{"LogNormalize"}. Details see \code{\link[Seurat]{NormalizeData}}}

\item{scale.factor}{Default: \code{10000}. Details see \code{\link[Seurat]{NormalizeData}}}

\item{selection.method}{Default: \code{"vst"}. Details see \code{\link[Seurat]{FindVariableFeatures}}}

\item{nfeatures}{Default: \code{2000}. Details see \code{\link[Seurat]{FindVariableFeatures}}}

\item{integrate.method}{Default: \code{Seurat::CCAIntegration}.
Details see \code{\link[Seurat]{IntegrateLayers}}}

\item{integrate.new.reduction}{Default: \code{"cca"}. Details see \code{\link[Seurat]{IntegrateLayers}}}

\item{dim.run.umap}{Default: \code{1:20}. Details see \code{\link[Seurat]{RunUMAP}}}

\item{dim.run.tsne}{Default: \code{1:20}. Details see \code{\link[Seurat]{RunTSNE}}}

\item{dim.find.neighbors}{Default: \code{1:20}. Details see \code{\link[Seurat]{FindNeighbors}}}

\item{resolution.find.clusters}{Default: \code{0.5}. Details see \code{\link[Seurat]{FindClusters}}}

\item{verbose}{Default: \code{FALSE}. Logical. Controls verbosity}
}
\value{
seurat object
}
\description{
Include: filtering cells by the percentage of "mitochondrial genes" (Default: no filter),
\code{"NormalizeData > FindVariableFeatures > ScaleData > RunPCA > IntegrateLayers(harmony)"}.
Users can also analyze the data themselves according to the standard workflow in \link{Seurat}.
}
\examples{
data(seurat.object)
seurat.object <- runSeurat(seurat.object)

}
