% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_analysisModule.R
\name{runTrajectory}
\alias{runTrajectory}
\title{Pseudotime Analysis of Single-Cell Data via runTrajectory(Monocle3-Based).}
\usage{
runTrajectory(seurat.object, method = "monocle3")
}
\arguments{
\item{seurat.object}{seurat object.}

\item{method}{Default: \code{"monocle3"}.}
}
\value{
seurat object, with cell trajectory result in misc$host$trajectory slot.
}
\description{
Pseudotime Analysis of Single-Cell Data via runTrajectory(Monocle3-Based).
}
\examples{
data("seurat.object")
# seurat.object <- runSeurat(seurat.object)
# seurat.object <- runTrajectory(seurat.object)

}
