% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat.object}
\alias{seurat.object}
\title{example data}
\format{
A Seurat object of five filtered 10X single cell samples.
}
\usage{
seurat.object
}
\description{
A Seurat object constructd by \code{\link{read_intraMisc}}
}
\keyword{datasets}
