% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{seurat2phyloseq}
\alias{seurat2phyloseq}
\title{Convert microbe seurat object to microbe phyloseq object.}
\usage{
seurat2phyloseq(microbe.seurat.object)
}
\arguments{
\item{microbe.seurat.object}{Microbe seurat object}
}
\value{
Microbe phyloseq object
}
\description{
Convert microbe seurat object to microbe phyloseq object.
}
\examples{
NULL

}
