% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_classes.R
\name{show,corrResult-method}
\alias{show,corrResult-method}
\title{Show Method for corrResult}
\usage{
\S4method{show}{corrResult}(object)
}
\arguments{
\item{object}{An instance of `corrResult`.}
}
\description{
Displays the correlation matrix, p-values matrix, and parameters used in the analysis.

This method provides a concise summary of the `corrResult` object, showing the first few rows and columns
of the correlation matrix and p-values matrix, as well as the parameters used for the analysis.
}
