% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMisc_baseFunction.R
\name{subset}
\alias{subset}
\title{reconstructed sbuset method}
\usage{
subset(object, ...)
}
\arguments{
\item{object}{S4 object.}

\item{...}{params of \code{subset} function from \strong{\code{SummarizedExperiment}}}
}
\value{
S4 object.
}
\description{
reconstructed sbuset method
}
\examples{
NULL

}
